Divided Iterator into ArrayAccess class
