+++ library/toolkit/Iterator.php
@@ -1,7 +1,7 @@
 <?php
 namespace Zerobase\Toolkit;
 
-abstract class Iterator
+abstract class Iterator extends ArrayAccess implements \Iterator
 {
     protected $data = array();
     protected $indexes = array();
@@ -35,29 +35,5 @@ abstract class Iterator
     public function valid()
     {
         return isset($this->data[$this->indexes[$this->position]]);
-    }
-
-    public function offsetExists($offset)
-    {
-        return isset($this->data[$offset]);
-    }
-
-    public function offsetSet($offset, $value)
-    {
-        if (is_null($offset)) {
-            $this->data[] = $value;
-        } else {
-            $this->data[$offset] = $value;
-        }
-    }
-
-    public function offsetGet($offset)
-    {
-        return $this->offsetExists($offset) ? $this->data[$offset] : null;
-    }
-
-    public function offsetUnset($offset)
-    {
-        unset($this->data[$offset]);
     }
 }
